LayoutTests/platform/wpe/imported/w3c/web-platform-tests/css/css-color-adjust/meta/meta-color-scheme-attribute-changes-expected.txt


FAIL Meta color-scheme initially 'dark'. assert_equals: expected (string) "dark" but got (undefined) undefined
FAIL Removed name attribute from meta color-scheme. assert_equals: expected (string) "normal" but got (undefined) undefined
FAIL Set meta name to color-scheme. assert_equals: expected (string) "dark" but got (undefined) undefined
FAIL Set content attribute of meta color-scheme to empty string. assert_equals: expected (string) "normal" but got (undefined) undefined
FAIL Set content attribute of meta color-scheme to an invalid value. assert_equals: expected (string) "normal" but got (undefined) undefined
FAIL Set content attribute of meta color-scheme to 'light'. assert_equals: expected (string) "light" but got (undefined) undefined
FAIL Removed the content attribute of meta color-scheme. assert_equals: expected (string) "normal" but got (undefined) undefined
